Move some utilites out
